more work on VEVENT toString
using map of component name as key and value as value
using joinging of streaming entrySet to make property strings

+++ src/main/java/jfxtras/labs/repeatagenda/scene/control/repeatagenda/VEventImpl.java
@@ -250,7 +250,7 @@ public class VEventImpl extends VEvent
     public WindowCloseType edit(
               LocalDateTime dateTimeOld
             , LocalDateTime dateTimeNew
-            , int durationInSecondsNew
+            , long durationInSecondsNew
             , VEventImpl vEventOld
             , Collection<Appointment> appointments
             , Collection<VEvent> vEvents
